👷Enables CDN for uploaded assets
Adds CloudFront CDN integration to serve uploaded assets.

Moves S3 bucket configuration and CDN resources to a separate file for better organization.
Adds an event-triggered function to move uploaded files to the CDN directory and update DynamoDB with the CDN URLs.
Adds the ability to delete objects to the upload bucket.