♻️ Basic changes to get function
Moved response generation into the handler for consistency.

Make query on the default table index.

Preprocess the object so that validation gets an instance of any subobjects such as the enum 'Status' instead of creating a new enum type object.

Change entity_type 'ORGANISATION' to 'ORG' for consistency with PK/SK.